requirements.txt
eel
pyqrcode
pyinstaller
pypng
autopep8
psycopg2-binary
SQLAlchemy
python-dotenv
